+++ src/pages/viewer/Channels.vue
@@ -16,13 +16,9 @@
           :label="active ? 'Channel' : 'SELECT CHANNEL'"
           dark hide-bottom-space dense color="white"
           :disable="!isNeedSelect"
-          :virtual-scroll-item-size="60"
-          :virtual-scroll-slice-size="6"
-          :virtual-scroll-sticky-size-start="48"
-          :virtual-scroll-sticky-size-end="needShowGetDeletedAction && tokenType === 1 ? 29 : 0"
           popup-content-class="items__popup"
           :popup-content-style="{height: `${((filteredItems.length > 6 ? 6 : filteredItems.length) * 60) + (needShowGetDeletedAction && tokenType === 1 ? 77 : 48)}px`}"
-          @filter="filterItems"
+          @filter="(filter, update) => filterItems('channels', filter, update)"
         >
           <div slot="before-options" class="bg-dark q-pa-xs select__filter">
             <q-input v-model="filter" @input="filter => $refs.itemSelect.filter(filter)" outlined hide-bottom-space rounded dense color="white" dark placeholder="Filter" autofocus>
@@ -271,14 +267,6 @@ export default {
   },
   methods: {
     ...mapActions(['getDeleted']),
-    filterItems (filter, update) {
-      if (this.isItemsInit) {
-        update()
-        return
-      }
-      const entity = 'channels'
-      this.itemsLoad(entity, update, this.active, () => { this.isItemsInit = true })
-    },
     viewDataHandler (content) {
       const message = content[content.length - 1]
       this.$emit('view-data', message)