+++ src/pages/viewer/Containers.vue
@@ -14,13 +14,9 @@
           :label="active ? 'Containers' : 'SELECT CONTAINER'"
           dark hide-bottom-space dense color="white"
           :disable="!isNeedSelect"
-          :virtual-scroll-item-size="48"
-          :virtual-scroll-slice-size="6"
-          :virtual-scroll-sticky-size-start="48"
-          :virtual-scroll-sticky-size-end="needShowGetDeletedAction && tokenType === 1 && false ? 29 : 0"
           popup-content-class="items__popup"
           :popup-content-style="{height: `${((filteredItems.length > 6 ? 6 : filteredItems.length) * 48) + (needShowGetDeletedAction && tokenType === 1 && false ? 77 : 48) + (filteredItems.length ? 0 : 4)}px`}"
-          @filter="filterItems"
+          @filter="(filter, update) => filterItems('containers', filter, update)"
         >
           <div slot="before-options" class="bg-dark q-pa-xs select__filter">
             <q-input v-model="filter" outlined hide-bottom-space rounded dense color="white" dark placeholder="Filter" @input="filter => $refs.itemSelect.filter(filter)" autofocus>
@@ -180,14 +176,6 @@ export default {
   },
   methods: {
     ...mapActions(['getDeleted']),
-    filterItems (filter, update) {
-      if (this.isItemsInit) {
-        update()
-        return
-      }
-      const entity = 'containers'
-      this.itemsLoad(entity, update, this.active, () => { this.isItemsInit = true })
-    },
     viewLogMessagesHandler (content) {
       this.$emit('view-log-message', content)
     },