+++ src/pages/viewer/Streams.vue
@@ -14,13 +14,9 @@
           dark hide-bottom-space dense color="white"
           :disable="!isNeedSelect"
           :hide-dropdown-icon="!isNeedSelect"
-          :virtual-scroll-item-size="48"
-          :virtual-scroll-slice-size="6"
-          :virtual-scroll-sticky-size-start="48"
-          :virtual-scroll-sticky-size-end="needShowGetDeletedAction && tokenType === 1 ? 29 : 0"
           popup-content-class="items__popup"
           :popup-content-style="{height: `${((filteredItems.length > 6 ? 6 : filteredItems.length) * 48) + (needShowGetDeletedAction && tokenType === 1 ? 77 : 48) + (filteredItems.length ? 0 : 4)}px`}"
-          @filter="filterItems"
+          @filter="(filter, update) => filterItems('streams', filter, update)"
         >
           <div slot="before-options" class="bg-dark q-pa-xs select__filter">
             <q-input v-model="filter" outlined hide-bottom-space rounded dense color="white" dark placeholder="Filter" @input="filter => $refs.itemSelect.filter(filter)" autofocus>
@@ -183,14 +179,6 @@ export default {
   },
   methods: {
     ...mapActions(['getDeleted']),
-    filterItems (filter, update) {
-      if (this.isItemsInit) {
-        update()
-        return
-      }
-      const entity = 'streams'
-      this.itemsLoad(entity, update, this.active, () => { this.isItemsInit = true })
-    },
     viewLogMessagesHandler (content) {
       this.$emit('view-log-message', content)
     },